requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
pydantic
email-validator
python-jose[cryptography]
passlib[bcrypt]
httpx
pydantic-settings
